Simulator kompilierbar.

+++ mock/badge_main_loop.c
@@ -32,8 +32,8 @@ void badge_main_loop(void) {
     menu[j] = menu_buf[j];
   }
 
-  size_t choice = 0;
-  size_t first_visible = 0;
+  uint8_t choice = 0;
+  uint8_t first_visible = 0;
 
   for(;;) {
     choice = badge_menu(menu, ARRAY_SIZE(menu), &first_visible, choice);
